requirements.txt
asgiref==3.8.1
black==23.3.0
click==8.1.7
colorama==0.4.6
Django==4.2.3
django-cors-headers==3.14.0
django-extensions==3.2.1
django-ratelimit==4.1.0
django-redis==5.4.0
django-throttle-requests==0.7.0
djangorestframework==3.14.0
flake8==6.0.0
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
platformdirs==4.3.6
psycopg2==2.9.6
pycodestyle==2.10.0
pyflakes==3.0.1
python-dotenv==1.0.0
python-memcached==1.62
pytz==2024.2
redis==5.2.0
shortuuid==1.0.9
sqlparse==0.5.1
tzdata==2024.2
